# specs/openapi.yaml
openapi: 3.1.0
info:
  title: relayauth API
  version: 1.0.0
  summary: Identity and authorization plane for the Agent Relay ecosystem.
  description: |-
    relayauth issues JWTs, manages identities, evaluates roles and policies,
    enforces scoped access, and exposes audit and admin APIs for the Agent Relay ecosystem.
...
servers:
  - url: http://localhost:8787
    description: Local development
  - url: https://api.relayauth.dev
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /.well-known/jwks.json:
    get:
      tags: [JWKS]
      summary: Get JSON Web Key Set
      description: Public signing keys for offline JWT validation.
      operationId: getJwks
      security: []
      responses:
        "200":
          description: JSON Web Key Set
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKSResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/tokens:
    post:
      tags: [Tokens]
      summary: Issue token pair
      description: Issue a new access and refresh token pair for an identity.
      operationId: createToken
      x-required-scopes: [relayauth:token:create]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/XWorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: Token pair created
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/tokens/refresh:
    post:
      tags: [Tokens]
      summary: Refresh token pair
      description: Exchange a refresh token for a new token pair.
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Refreshed token pair
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/tokens/revoke:
    post:
      tags: [Tokens]
      summary: Revoke token or identity tokens
      description: Revoke a specific token, a session, or all tokens for an identity.
      operationId: revokeToken
      x-required-scopes: [relayauth:token:manage]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeTokenRequest"
      responses:
        "204":
          description: Token(s) revoked
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/identities:
    post:
      tags: [Identities]
      summary: Create identity
      description: Create a new agent, human, or service identity.
      operationId: createIdentity
      x-required-scopes: [relayauth:identity:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdentityRequest"
      responses:
        "201":
          description: Identity created
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Identities]
      summary: List identities
      description: List identities with optional filters for status, type, and workspace.
      operationId: listIdentities
      x-required-scopes: [relayauth:identity:read]
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/IdentityStatus"
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/IdentityType"
        - in: query
          name: workspaceId
          schema:
            type: string
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Identities page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIdentityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/identities/{identityId}:
    get:
      tags: [Identities]
      summary: Get identity
      description: Retrieve a single identity by ID.
      operationId: getIdentity
      x-required-scopes: [relayauth:identity:read]
      parameters:
        - $ref: "#/components/parameters/IdentityId"
      responses:
        "200":
          description: Identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Identities]
      summary: Update identity
      description: Update mutable fields on an identity.
      operationId: updateIdentity
      x-required-scopes: [relayauth:identity:manage]
      parameters:
        - $ref: "#/components/parameters/IdentityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIdentityRequest"
      responses:
        "200":
          description: Updated identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Identities]
      summary: Delete identity
      description: Hard delete intended only for identities without durable audit history.
      operationId: deleteIdentity
      x-required-scopes: [relayauth:identity:manage]
      parameters:
        - $ref: "#/components/parameters/IdentityId"
      responses:
        "204":
          description: Identity deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/identities/{identityId}/suspend:
    post:
      tags: [Identities]
      summary: Suspend identity
      description: Temporarily deactivate an identity. Can be reactivated later.
      operationId: suspendIdentity
      x-required-scopes: [relayauth:identity:manage]
      parameters:
        - $ref: "#/components/parameters/IdentityId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendIdentityRequest"
      responses:
        "200":
          description: Suspended identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/identities/{identityId}/reactivate:
    post:
      tags: [Identities]
      summary: Reactivate identity
      description: Reactivate a suspended identity. Retired identities cannot be reactivated.
      operationId: reactivateIdentity
      x-required-scopes: [relayauth:identity:manage]
      parameters:
        - $ref: "#/components/parameters/IdentityId"
      responses:
        "200":
          description: Reactivated identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/identities/{identityId}/retire:
    post:
      tags: [Identities]
      summary: Retire identity
      description: Permanently deactivate an identity and revoke all active tokens.
      operationId: retireIdentity
      x-required-scopes: [relayauth:identity:manage]
      parameters:
        - $ref: "#/components/parameters/IdentityId"
      responses:
        "200":
          description: Retired identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/scopes/templates:
    get:
      tags: [Scopes]
      summary: List scope templates
      operationId: listScopeTemplates
      x-required-scopes: [relayauth:scope:read]
      responses:
        "200":
          description: Predefined scope templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeTemplatesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by relayauth.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Organization API key used for token issuance.
...
  parameters:
...
    XWorkspaceId:
      in: header
      name: X-Workspace-Id
      required: false
      schema:
        type: string
      description: Optional workspace context override.
  responses:
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    JWKSResponse:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/Jwk"
...
    CreateIdentityRequest:
      $ref: "#/components/schemas/CreateIdentityInput"
...
    CreateTokenRequest:
      type: object
      required: [identityId]
      properties:
        identityId:
          type: string
        scopes:
          type: array
          items:
            type: string
        workspaceId:
          type: string
        ttl:
          type: string
          default: 1h
        audience:
          type: array
          items:
            type: string
...